init: Add androidboot.bootmedia variable
Init will now try to execute an init.$(androidboot.bootmedia).rc
prior to the main init.rc. We use this to separate out mounting
partitions so that we can support multiple bootmedia types,
since they will map to different block device nodes.

If the value is not passed in via kernel command line, it defaults
to sdcard. If the rc file doesn't exist, boot will skip it.

We also insert a NULL check in cur_command->func.

Change-Id: Ic4e417a1d47cb9ad815f1599441c5870ec85d22e