init: Make control-C work on the console
Job control (e.g., control-C) doesn't work for the Android
console shell.  This change sets the controlling tty for the
console service so that job control works.

Design:  When a service is launched with the "console"
attribute, issue a TIOCSCTTY ioctl to its stdin
file-descriptor to set it as the controlling tty.  The ioctl
is issued with the "try" instead of "steal" option in case
there is already ownership of the controlling tty.
Presumably only the console shell has the "console"
attribute, so this should not be a concern in any case.

Change-Id: I281c1beb0ae1d4ba17dc337a2c753459c0bd217b
Signed-off-by: Paul Drews <paul.drews@intel.com>