charger: abstract to library, policy adjustments
The core charge graphics drawing logic has been moved to
a static libcharger library so that it can be used elsewhere
to report battery charging.

The main charger application is now in main.c. Many configuration
options which were preprocessor defines are now parameters to
the charger_run() function, which is the library entry point.
The policy for what to do when the charger mechanism exits with
shutdown or proceed status is left to the caller.

The charger has a new parameter min_charge to stay in the charger
until the battery level reaches a specific percentage.

Connecting/disconnecting a power source is now sufficient stimuli
to restart the animation if it had reached the cycle limit.

Some warnings have been cleaned up and -Wall -Werror enforced
in the compilation of these modules.

Graphical resources are now freed when leaving charger_run().

Issue: ANDROID-1132
Change-Id: Iac06ea0c51f2ff60c8630f79cfda3a4cee5fd5f5
Depends-Change-Id: Id4a9eda919ad57fbd0c465c358427e5de81caabd
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>