Let ueventd auto-load kernel modules
Implements the functionality for ueventd to auto-load kernel
modules when uevents are triggered.

Since /system may not be mounted when uevents are fired,
a deferred loading mechanism is implemented. Once mapping of
module and alias is available, these modules are then loaded.

Modules can also be blacklisted so they will not be loaded
automatically. One example would be the Wifi driver, as
Android's has to control its loading and unloading.

Issue: AXIA-170
Change-Id: I755b0d276db0f90dd0e578cd7aacb320ed5a3082
Depends-Change-Id: I74653e64e079af3cde483b7f4b9110456ce18fc8
Signed-off-by: Daniel Leung <daniel.leung@intel.com>