Deferred module loading may cause ueventd to segfault
When going through the list for deferred module loading,
the list nodes are freed but the for loop may need to access
it for pointer to next node. This will cause segfault in
some situations where the freed memory got allocated by others
immediately.

Implement a list_for_each_safe() marco that is safe against
removal of list entries. These entries can be freed (if malloc)
inside the node without causing segfault.

For: AXIA-589
Change-Id: I43590726f9df1203c945a9388607e8733feef1ac
Signed-off-by: Daniel Leung <daniel.leung@intel.com>