init: remove logo code
This code is legacy and should never need to be used. The bootloader
should put an image on the screen and the kernel shouldn't reconfigure
the display controller until userland is ready to draw something,
such as the boot animation, charger, recovery console, etc.

Instead of hard-coding tty0, the "A N D R O I D" text is written out
to whatever the active console is.

Change-Id: I5fd21413f721d2ba6966d34d632d8fc3913ca073
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>