Handle kernel requests of modules in ueventd
Drivers in kernel can request modules by launching a program in
user space, the program's path by default is "/sbin/modprobe".

Because Android system has no modprobe and ueventd is the only
program handling the module aliases so far, This patch provides a
cheap approach to handle kernel's requests in ueventd executable.

e.g. sound sub-system in kernel can get codec modules loaded based
on probing of codec chips on device in this way.

NOTE:
This feature is intended to be used by root (uid 0) only, and it
is NOT an equivalent of GNU modprobe.

Some requests from kernel could not be handled if they come to
user space before file system for alias files is mounted.

For: AXIA-577
Change-Id: I7eefe17fd106843e6905b6bee9377f9bb662db99
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>