adb: do not force group/world permissions when pushing files
When adb pushes files, it copies the user permission to group and
world (e.g. file with mode 0644 becomes 0666 after pushed). This
causes problems with init as it ignores files with writable
permission for group and world. This change removes the code
to copy permissions so that the destination files retain the
original permission.

Issue: AXIA-630
Change-Id: Idbabc7339fbc97781bdd8094c43b4eed094226e1
Signed-off-by: Daniel Leung <daniel.leung@intel.com>