Renaming getline to fs_getline for fs_mgr
fs_mgr defines its own version of getline and uses it
internally. This leads to a build error if getline is
also defined in bionic, since fs_mgr will see readline
as defined internally.

Renaming getline in fs_mgr to a local name (fs_getline)
so that there will no longer be any conflicts.

This is needed it we want to add getline in bionic.

For: AIA-452
Change-Id: Ifae2f2e991cca7935e140b692fbe03cee2d917fc
Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>