Add aklog tool for kernel log
This small tool provides some new features, compared with existing
dmesg and /proc/kmsg approaches to get RAW kernel log.

() It can be run in a service style to output log continuously
   until it's killed. (e.g. launch it as a service to collect log
   at run time.

() It can be run once to output all logs and then quit. (e.g. run
   it in this way to get all logs after a reboot.)

() It supports to output logs to console or a file, the later
   option works with service manner to keep logs after a reboot.

For: AXIA-765
Change-Id: I006364ddb4627a3fe5a99026913b1e291e83e9cc
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>