init.trace.rc: Key off of ro.debuggable, mount debugfs
Currently, if you want to use debugfs for tracing, you need to mount it
before the 'on boot' block in this file is processed, else the permissions
will be wrong. This means enabling debugfs is an all-or-nothing proposition,
it's not possible to do something in your init.<platform>.rc like:

    on property:ro.debuggable=1
        mount debugfs debugsfs /sys/kernel/debug

Since 'on property' triggers happen much later in the boot process.

The board configurations in AOSP all put the call to mount debugfs
in the early-init or boot blocks of theiir init.$(TARGET_PRODUCT).rc. This
is a security issue, debugfs should not be mounted in a production
configuration:

http://lwn.net/Articles/429328/

What this change does is to only enable debugfs and set the tracing
permissions in debug-enabled builds, and automatically mount debugfs
in those configurations.

Change-Id: I9e9734c2c453b07afbb6d01db80c591ef1bf816d
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>