init: leave ro.serialno unset if not provided in command line
Some systems may use bootloaders which can't modify the kernel
command line to provide the unit serial number, and this must
be supplied later. In order to allow this we need to not set
ro.serialno to an empty value, otherwise it's impossible to set it
to the real serial number later in the boot process.

Change-Id: Ieeb085c19437286400729aa00d2463ac9f8cec84
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>