libcutils: Add BCB writing support
Our Linux patch to update the BCB in /misc to do a oneshot selection
of the next boot target was rejected upstream; they do not want the
REBOOT2 argument used in this way. So for devices which specify a device
node using 'androidboot.bcb_device=XXX' in the command line, update the
BCB there upon rebooting in userspace.

Change-Id: I5750d70958a195b91f10e18820294abe85e18f0c
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>