init: create by-name device links for marked GPT partitions
Create device node symlinks by name if the uevent for a block
device has a name in it, and the name passed in matches the
prefix for the installation ID. This prevents collisions with
multiple Android installations on the same system. The created
symlinks do not have the prefix in them.

If the installation ID isn't set (through ro.boot.install_id)
this mechanism doesn't do anything. There is a possibility that
specially crafted removable storage could alias internal storage
through this mechanism, so locked-down devices won't want to use
this.

Change-Id: I04782691d4e5ff0db6f68d8d9345a8585cfe5f86
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>