always try to create by-name links based on install id
Previously, if a block device was detected as a platform device,
the by-name logic which uses install id was skipped in favor of
the older logic which includes the bus in the path name. This
resulted in the installation of Android failing on these devices
since the expected devices nodes don't show up after the disk is
partitioned.

Now always try to use the install id logic, and fall back to the
older mechanism only if it doesn't produce any symlinks.

Change-Id: Idc096eb6425af478ddca7fe4ba9c0fe24b922d6e
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>